feat: export GenerateCompatibilityResults function and add Encoding.RegisterProvider
